kntl

+++ admin/paket_tour.php
@@ -51,7 +51,7 @@
                   d.nama_destinasi, d.id_destinasi,
                   mt.kapasitas, mt.image_travell
                   FROM paket_tour AS pt
-                  JOIN destinasi AS d ON pt.id_destinasi = d.id_destinasi
+                  JOIN destinasi AS d ON pt.id_paket_tour = d.id_paket_tour
                   JOIN mobil_travel AS mt ON pt.kode_travell = mt.kode_travell";
       $query = mysqli_query($kon, $select);
       echo "
@@ -64,19 +64,30 @@
                 <th>Description</th>
                 <th>Booking</th>
           ";
-
+      $id = 0;
+      $id_des = null;
+      $nama_des = null;
       while ($row = mysqli_fetch_assoc($query)) {
-        echo "
-              <tr>
-                <td align='center'>".$row['nama_paket']."</td>
-                <td align='center'><a href='detail_destinasi.php?id_destinasi=".$row['id_destinasi']."'>".$row['nama_destinasi']."</a></td>
-                <td align='center'><img src='../pict/".$row['image_travell']."' width='150' height='150'></td>
-                <td align='center'>".$row['kapasitas']." Penumpang</td>
-                <td align='center'>Rp. ".$row['harga_paket']."</td>
-                <td align='center'>".$row['description']."</td>
-                <td align='center'><a href='edit_paket_tour.php?id_paket=".$row['id_paket_tour']."'>Edit</a><br/><a href='hapus_paket_tour.php?id_paket=".$row['id_paket_tour']."'>Hapus</a></td>
-              </tr>
-            ";
+        if($id == $row['id_paket_tour']) {
+          $nama_des = $row['nama_destinasi'];
+          $id_des = $row['id_destinasi'];
+        } else {
+          echo "
+          <tr>
+            <td align='center'>".$row['nama_paket']."</td>
+            <td align='center'>
+              <a href='detail_destinasi.php?id_destinasi=".$row['id_destinasi']."'>".$row['nama_destinasi'].",</a>
+              <a href='detail_destinasi.php?id_destinasi=".$id_des."'>".$nama_des."</a>
+            </td>
+            <td align='center'><img src='../pict/".$row['image_travell']."' width='150' height='150'></td>
+            <td align='center'>".$row['kapasitas']." Penumpang</td>
+            <td align='center'>Rp. ".$row['harga_paket']."</td>
+            <td align='center'>".$row['description']."</td>
+            <td align='center'><a href='edit_paket_tour.php?id_paket=".$row['id_paket_tour']."'>Edit</a><br/><a href='hapus_paket_tour.php?id_paket=".$row['id_paket_tour']."'>Hapus</a></td>
+          </tr>
+        ";
+        $id = $row['id_paket_tour'];
+        }
       }
       echo "</table>";
   ?>
